requirements.txt
flask = 2.3.3
my-sql-connector-python = 8.1.0